added signature verification
